rearrange files, watchlist and analyst services
